[fix] Fix directory+names (#24)
