--- trunk/as0005409/task_01/src/Lab.1.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29905.134
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Lab.1", "Lab.1\Lab.1.vcxproj", "{2DD87175-952F-4424-8597-9460BF6485F6}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2DD87175-952F-4424-8597-9460BF6485F6}.Debug|x64.ActiveCfg = Debug|x64
-		{2DD87175-952F-4424-8597-9460BF6485F6}.Debug|x64.Build.0 = Debug|x64
-		{2DD87175-952F-4424-8597-9460BF6485F6}.Debug|x86.ActiveCfg = Debug|Win32
-		{2DD87175-952F-4424-8597-9460BF6485F6}.Debug|x86.Build.0 = Debug|Win32
-		{2DD87175-952F-4424-8597-9460BF6485F6}.Release|x64.ActiveCfg = Release|x64
-		{2DD87175-952F-4424-8597-9460BF6485F6}.Release|x64.Build.0 = Release|x64
-		{2DD87175-952F-4424-8597-9460BF6485F6}.Release|x86.ActiveCfg = Release|Win32
-		{2DD87175-952F-4424-8597-9460BF6485F6}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {AEA66609-8AAA-444A-9C87-F4808BA8BA1A}
-	EndGlobalSection
-EndGlobal
