--- trunk/as0005414/task_01/src/ConsoleApplication1.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30413.136
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ConsoleApplication1", "ConsoleApplication1.vcxproj", "{484B7BBA-E5CC-4868-934D-971F7B162738}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{484B7BBA-E5CC-4868-934D-971F7B162738}.Debug|x64.ActiveCfg = Debug|x64
-		{484B7BBA-E5CC-4868-934D-971F7B162738}.Debug|x64.Build.0 = Debug|x64
-		{484B7BBA-E5CC-4868-934D-971F7B162738}.Debug|x86.ActiveCfg = Debug|Win32
-		{484B7BBA-E5CC-4868-934D-971F7B162738}.Debug|x86.Build.0 = Debug|Win32
-		{484B7BBA-E5CC-4868-934D-971F7B162738}.Release|x64.ActiveCfg = Release|x64
-		{484B7BBA-E5CC-4868-934D-971F7B162738}.Release|x64.Build.0 = Release|x64
-		{484B7BBA-E5CC-4868-934D-971F7B162738}.Release|x86.ActiveCfg = Release|Win32
-		{484B7BBA-E5CC-4868-934D-971F7B162738}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {C50370FB-EF69-4056-8298-E9E1BA350806}
-	EndGlobalSection
-EndGlobal
