+++ trunk/as005402/task_01/src/source.cpp
