+++ trunk/as005409/task_01/doc/img/equation.png
