+++ trunk/as005412/task_01/doc/image/sub_eq.png
