+++ trunk/as005412/task_01/src/lab1/lab1.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30517.126
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "lab1", "lab1.vcxproj", "{8EE3C1CB-7649-466A-BBF4-373793190BC3}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{8EE3C1CB-7649-466A-BBF4-373793190BC3}.Debug|x64.ActiveCfg = Debug|x64
+		{8EE3C1CB-7649-466A-BBF4-373793190BC3}.Debug|x64.Build.0 = Debug|x64
+		{8EE3C1CB-7649-466A-BBF4-373793190BC3}.Debug|x86.ActiveCfg = Debug|Win32
+		{8EE3C1CB-7649-466A-BBF4-373793190BC3}.Debug|x86.Build.0 = Debug|Win32
+		{8EE3C1CB-7649-466A-BBF4-373793190BC3}.Release|x64.ActiveCfg = Release|x64
+		{8EE3C1CB-7649-466A-BBF4-373793190BC3}.Release|x64.Build.0 = Release|x64
+		{8EE3C1CB-7649-466A-BBF4-373793190BC3}.Release|x86.ActiveCfg = Release|Win32
+		{8EE3C1CB-7649-466A-BBF4-373793190BC3}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {1A487937-D881-41EF-BA3E-74D0C1B152C2}
+	EndGlobalSection
+EndGlobal
